+++ sourceCode/dynamics/RigidBodyContainerDyn_base.h
@@ -29,44 +29,44 @@ public:
     CRigidBodyContainerDyn_base();
     virtual ~CRigidBodyContainerDyn_base();
 
-    virtual std::string init(const float floatParams[20],const int intParams[20]);
+    virtual std::string init(const double floatParams[20],const int intParams[20]);
 
     virtual std::string getEngineInfo() const;
     virtual void serializeDynamicContent(const std::string& filenameAndPath,int maxSerializeBufferSize);
 
-    virtual void handleDynamics(float dt,float simulationTime);
+    virtual void handleDynamics(double dt,double simulationTime);
     virtual bool isDynamicContentAvailable();
-    bool getContactForce(int dynamicPass,int objectHandle,int index,int objectHandles[2],float* contactInfo);
+    bool getContactForce(int dynamicPass,int objectHandle,int index,int objectHandles[2],double* contactInfo);
     int getDynamicsCalculationPasses() const;
-    float getSimulationTime() const;
-    float* getContactPoints(int* cnt);
+    double getSimulationTime() const;
+    double* getContactPoints(int* cnt);
 
     static void setDynWorld(CRigidBodyContainerDyn* dynWorld);
     static CRigidBodyContainerDyn* getDynWorld();
     CParticleObjectContainer_base* getParticleCont();
 
-    float getPositionScalingFactorDyn() const;
-    float getLinearVelocityScalingFactorDyn() const;
-    float getMassScalingFactorDyn() const;
-    float getMasslessInertiaScalingFactorDyn() const;
-    float getForceScalingFactorDyn() const;
-    float getTorqueScalingFactorDyn() const;
-    float getGravityScalingFactorDyn() const;
-    float getDynamicsInternalTimeStep() const;
+    double getPositionScalingFactorDyn() const;
+    double getLinearVelocityScalingFactorDyn() const;
+    double getMassScalingFactorDyn() const;
+    double getMasslessInertiaScalingFactorDyn() const;
+    double getForceScalingFactorDyn() const;
+    double getTorqueScalingFactorDyn() const;
+    double getGravityScalingFactorDyn() const;
+    double getDynamicsInternalTimeStep() const;
     int getDynamicParticlesIdStart() const;
 
     static bool isJointInDynamicMode(CXSceneObject* joint);
 
 protected:
     virtual void _applyGravity();
-    virtual void _stepDynamics(float dt,int pass);
+    virtual void _stepDynamics(double dt,int pass);
     virtual void _createDependenciesBetweenJoints();
     virtual void _removeDependenciesBetweenJoints(CConstraintDyn* theInvolvedConstraint);
     virtual bool _updateWorldFromCoppeliaSim();
 
-    void _reportRigidBodyStatesToCoppeliaSim(float simulationTime);
-    void _reportConstraintStatesToCoppeliaSim(float simulationTime,int currentPass,int totalPasses);
-    virtual void _reportWorldToCoppeliaSim(float simulationTime,int currentPass,int totalPasses);
+    void _reportRigidBodyStatesToCoppeliaSim(double simulationTime);
+    void _reportConstraintStatesToCoppeliaSim(double simulationTime,int currentPass,int totalPasses);
+    virtual void _reportWorldToCoppeliaSim(double simulationTime,int currentPass,int totalPasses);
     void _handleAdditionalForcesAndTorques();
     void _clearAdditionalForcesAndTorques();
     CRigidBodyDyn* _getRigidBodyFromObjectHandle(int shapeHandle);
@@ -92,8 +92,8 @@ protected:
     void _updateHybridJointTargetPositions_old();
 
     // Following 3 go in pair:
-    void _handleKinematicBodies_init(float dt);
-    void _handleKinematicBodies_step(float t,float cumulatedTimeStep);
+    void _handleKinematicBodies_init(double dt);
+    void _handleKinematicBodies_step(double t,double cumulatedTimeStep);
     void _handleKinematicBodies_end();
 
     static CRigidBodyContainerDyn* _dynWorld;
@@ -102,21 +102,21 @@ protected:
     std::map<int,CConstraintDyn*> _allConstraints;
     CParticleObjectContainer_base* _particleCont;
 
-    std::vector<float> _contactPoints;
+    std::vector<double> _contactPoints;
     std::vector<SContactInfo> _contactInfo; // Not same as above!
 
     int _dynamicsCalculationPasses;
-    float _simulationTime;
+    double _simulationTime;
     int _engine;
     int _engineVersion;
-    float _positionScalingFactorDyn;
-    float _linearVelocityScalingFactorDyn;
-    float _massScalingFactorDyn;
-    float _masslessInertiaScalingFactorDyn;
-    float _forceScalingFactorDyn;
-    float _torqueScalingFactorDyn;
-    float _gravityScalingFactorDyn;
-    float _dynamicActivityRange;
-    float _dynamicsInternalStepSize;
+    double _positionScalingFactorDyn;
+    double _linearVelocityScalingFactorDyn;
+    double _massScalingFactorDyn;
+    double _masslessInertiaScalingFactorDyn;
+    double _forceScalingFactorDyn;
+    double _torqueScalingFactorDyn;
+    double _gravityScalingFactorDyn;
+    double _dynamicActivityRange;
+    double _dynamicsInternalStepSize;
     int _dynamicParticlesIdStart;
 };
